Fix worker tests
